appium_tests/requirements.txt
Appium-Python-Client==3.1.0
pytest==8.2.0
pytest-html==4.1.1
pytest-xdist==3.5.0
pytest-ordering==0.6
allure-pytest==2.13.2
requests==2.31.0
PyYAML==6.0.1
python-dotenv==1.0.1
Faker==24.0.0
